bunch of changes: - MVVM: - - Model: - - - DisciplineType: added enum for better understanding type of DisciplineType; - - - Discipline: changed `int LessonType` to `DisciplineType LessonType`; - - - Lesson: changed types of Number and SubgroupCount from int to byte to match API; - - - TimeTable: changed `DateTime Date` to `DateTimeOffset Date` to match API; - - View: - - - TimeTableView: created view for student diary; - - ViewModel: - - - TimeTableViewModel: added binding values from MrsuStorageService; - Resources: - - Styles: - - - DarkTheme: added new color: SubDark #252629 - - - LightTheme: added new color: SubLight #b2b3b7, changed SecondaryLight from #b2b3b7 to #a0a1a5; - Services: - - IMrsuApiService: changed GetTimeTable parameter date from type DateTime to DateOnly and return value from Task<StudentTimeTable> to Task<List<StudentTimeTable>>; - - MrsuApiService: changed GetTimeTable parameters to match IMrsuApiService and the way getting data from API; - - MrsuStorageService: added Diary Dictionary to store any student's TimeTables and override GetTimeTable to prevent always getting data from API (now it should take from local storage aka Diary) common.ruleset: changed action of rule SA1602 from Error to None. And other small changes.

+++ MRSUMobile/MVVM/Model/Lesson.cs
@@ -2,9 +2,9 @@
 {
     public class Lesson
     {
-        public int Number { get; set; }
+        public byte Number { get; set; }
 
-        public int SubgroupCount { get; set; }
+        public byte SubgroupCount { get; set; }
 
         public List<Discipline> Disciplines { get; set; }
     }
